refact: widgets
