--- pyinstaller-command.json
@@ -1,73 +1,0 @@
-{
- "version": "auto-py-to-exe-configuration_v1",
- "pyinstallerOptions": [
-  {
-   "optionDest": "noconfirm",
-   "value": true
-  },
-  {
-   "optionDest": "filenames",
-   "value": "D:/STC_APP/improvements/security-layer/datadecryption_poetry/main.py"
-  },
-  {
-   "optionDest": "onefile",
-   "value": false
-  },
-  {
-   "optionDest": "console",
-   "value": true
-  },
-  {
-   "optionDest": "icon_file",
-   "value": "D:/STC_APP/improvements/security-layer/datadecryption_poetry/resources/stc_logo.ico"
-  },
-  {
-   "optionDest": "name",
-   "value": "DataGenerator V 0.1"
-  },
-  {
-   "optionDest": "clean_build",
-   "value": false
-  },
-  {
-   "optionDest": "loglevel",
-   "value": "ERROR"
-  },
-  {
-   "optionDest": "strip",
-   "value": false
-  },
-  {
-   "optionDest": "noupx",
-   "value": false
-  },
-  {
-   "optionDest": "disable_windowed_traceback",
-   "value": false
-  },
-  {
-   "optionDest": "version_file",
-   "value": "C:/Users/hamza.qureshi/Desktop/STC_APP/improvements/security-layer/datagen-v39-vscode-v2 - (xDataBase)/version_info.txt"
-  },
-  {
-   "optionDest": "uac_admin",
-   "value": false
-  },
-  {
-   "optionDest": "uac_uiaccess",
-   "value": false
-  },
-  {
-   "optionDest": "argv_emulation",
-   "value": false
-  },
-  {
-   "optionDest": "bootloader_ignore_signals",
-   "value": false
-  }
- ],
- "nonPyinstallerOptions": {
-  "increaseRecursionLimit": true,
-  "manualArguments": ""
- }
-}
