+++ src/gui/widgets/__init__.py
@@ -1,0 +1,2 @@
+from .extras import *
+from .table import *
